moved entity manipulation functionality to
