+++ app/src/App.css
@@ -22,7 +22,7 @@ body {
  * this is a hack around this issue  https://github.com/palantir/blueprint/issues/3515
  * it enables users to interact with canvas while drawer is open
 */
-.bp3-overlay {
+.bp3-overlay-container {
   pointer-events: none;
 }
 .bp3-drawer {
