+++ lib/package.json
@@ -79,9 +79,13 @@
   "peerDependencies": {
     "@alephdata/followthemoney": "^1.11.20",
     "@blueprintjs/core": "^3.15.1",
-    "d3": "^5.9.2",
+    "@blueprintjs/datetime": "^3.9.0",
+    "@blueprintjs/select": "^3.8.0",
     "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-draggable": "^3.3.0"
+    "react-dom": "^16.8.6"
+  },
+  "dependencies": {
+    "react-draggable": "^3.3.0",
+    "d3": "^5.9.2"
   }
 }
