+++ lib/rollup.config.ts
@@ -5,14 +5,15 @@ import typescript from 'rollup-plugin-typescript2'
 import json from 'rollup-plugin-json'
 
 const pkg = require('./package.json');
+const external = pkg['peerDependencies'] && Object.keys(pkg['peerDependencies']);
 
 export default {
   input: `src/index.ts`,
   output: [
-    { file: pkg.main, name:'vislib', format: 'umd', sourcemap: true },
-    { file: pkg.module, format: 'es', sourcemap: true },
+    {file: pkg.main, name: 'vislib', format: 'umd', sourcemap: true},
+    {file: pkg.module, format: 'es', sourcemap: true},
   ],
-  external: [ 'react', 'react-dom', '@alephdata/followthemoney', 'd3' ],
+  external,
   watch: {
     include: 'src/**',
   },
@@ -21,13 +22,13 @@ export default {
     // Allow node_modules resolution, so you can use 'external' to control
     // which external modules to include in the bundle
     // https://github.com/rollup/rollup-plugin-node-resolve#usage
-    resolve({ browser:true,}),
+    resolve({browser: true,}),
     // Compile TypeScript files
-    typescript({ useTsconfigDeclarationDir : true, objectHashIgnoreUnknownHack:true}),
+    typescript({useTsconfigDeclarationDir: true, objectHashIgnoreUnknownHack: true}),
     // Allow bundling cjs modules (unlike webpack, rollup doesn't understand cjs)
     commonjs({
-      namedExports:{
-        "react-draggable":['DraggableCore', 'DraggableEvent']
+      namedExports: {
+        "react-draggable": ['DraggableCore', 'DraggableEvent']
       }
     }),
     // Resolve source maps to the original source
