+++ lib/src/ftm/type/TextType.tsx
