+++ lib/src/ftm/type/common.tsx
